Extract HandlePosition enum into its own file
Decouples state types from view code — CanvasSelectionState
depended on HandlePosition which lived in SelectionOverlay.swift (a View file).
